Refactor the post data structure
+ Refactor the imgurls field in post data
The imgurls is now refactored to images.
It's an array which contains image object {url, width, height}
The image width and height is fetched by server in advance.
+ The postDate field in post data is no longer used.
+ Add getImageSize() to util.js
This function will fetch the width and height for a specified
image url.
+ Fix database service testing issue
Fix a database service testing assertion to match the change
of post data structure.