Refactor the return result when read post
The original design of the get post route and readPost function
will return an empty object with 200 status code when there is
no any post exist.

It is a little confused when trying to recognise the query result.
It now return a 404 status code with a simple message as a clear
result.

This commit applies:

+ Change the status code to 404 with a simple message when there's
no any result in query
+ Add eslint to help static analysis
+ Refactor the variable naming
+ Refactor code structures